Fix for RPM upgrade path of `intel-tee-pcs-client-tool`
Added upgrade guard in RPM `%preun` section to skip cleanup during package upgrades.
Moved wrapper script creation from `%post` to `%posttrans` in RPM spec.
Changed venv directory from `venv` to `.venv` for both RPM and DEB packages.

---------

+++ tools/PcsClientTool/installer/linux/rpm/intel-tee-pcs-client-tool/intel-tee-pcs-client-tool.spec
@@ -4,7 +4,7 @@
 #
 
 %define _install_path @install_path@
-%define _venv_dir @install_path@/venv
+%define _venv_dir @install_path@/.venv
 %define _main_script_name @main_script_name@
 %define _wrapper_script_name @pkg_wrapper_script_name@
 %define _license_file COPYING
@@ -103,6 +103,10 @@ function on_pip_error() {
     exit ${error_code}
 }
 
+if [ -d "%{_venv_dir}" ]; then
+    echo "Removing old virtual environment at %{_venv_dir}"
+    rm -rf "%{_venv_dir}"
+fi
 echo "Creating virtual environment..."
 %{python_interpreter} -m venv "%{_venv_dir}"
 
@@ -127,6 +131,11 @@ trap - ERR
 
 echo "Virtual environment setup complete."
 
+
+%posttrans
+# Note: due to a broken package in 1.24 release (which deletes these files unconditionally in %preun which runs after %post (new),
+#        we're installing them at the end of transaction).
+#       If this is ever deemed too late, it can be moved back inside %post, but only AFTER 1.24 release is out of support.
 echo -n "Creating wrapper script at '%{_install_path}/%{_wrapper_script_name}' and linking to /usr/local/bin/..."
 
 cat <<EOF > "%{_install_path}/%{_wrapper_script_name}"
@@ -143,6 +152,20 @@ echo "Installation successful. Run \`%{_wrapper_script_name}\` to start using th
 
 
 %preun
+# Note this cleanup part runs also on upgrade path (after the %post of the new package), so that we always start without cached files in a fresh venv, similar to DEB installer logic
+#
+if [ -d "%{_install_path}" ]; then
+    #remove any __pycache__ remnant dirs which may have been created in "wrong" location if user launched the script w/o wrapper script
+    find "%{_install_path}" -type d -name "__pycache__" -exec rm -rf {} +
+else
+    echo "Installation path %{_install_path} does not exist. Skipping __pycache__ removal."
+fi
+
+if [ $1 -ne 0 ]; then
+    # On upgrade path, we stop the cleanup here
+    exit 0
+fi
+
 if [ -d "%{_venv_dir}" ]; then
     echo "Removing virtual environment at %{_venv_dir}"
     rm -rf "%{_venv_dir}"
@@ -155,13 +178,6 @@ if [ -f "%{_install_path}/%{_wrapper_script_name}" ]; then
     rm -f "%{_install_path}/%{_wrapper_script_name}"
 else
     echo "No wrapper script found at %{_install_path}/%{_wrapper_script_name}. Skipping its removal."
-fi
-
-if [ -d "%{_install_path}" ]; then
-    #remove any __pycache__ remnant dirs which may have been created in "wrong" location if user launched the script w/o wrapper script
-    find "%{_install_path}" -type d -name "__pycache__" -exec rm -rf {} +
-else
-    echo "Installation directory %{_install_path} does not exist. Skipping __pycache__ removal."
 fi
 
 # remove the symlink from /usr/local/bin
@@ -177,6 +193,9 @@ if [ $1 == 0 -a -x %{_install_path}/cleanup.sh ]; then %{_install_path}/cleanup.
 
 
 %changelog
-* Wed Oct 01 2025 Intel Confidential Computing Team <confidential.computing@intel.com> - @version@
+* @date@ Intel Confidential Computing Team <confidential.computing@intel.com> - @version@-1
+- Installer: moved Python Virtual Environment to '.venv' and fixed version upgrade path (%preun upgrade guard)
+
+* Wed Oct 01 2025 Intel Confidential Computing Team <confidential.computing@intel.com> - 1.24.100.2-1
 - Initial packaged release of pcsclient.py. 
   Follows the former (all-in-one) pccsadmin.py tool split into (PCCS-specific) pccsadmin.py and (generic) pcsclient.py
